Crate blogger3

Features

Handle the following Resources with ease from the central hub ...

blog user infos (get)

blogs (get, getbyurl and listbyuser)

comments (approve, delete, get, list, listbyblog, markasspam and removecontent)

page views (get)

pages (delete, get, insert, list, patch, publish, revert and update)

post user infos (get and list)

posts (delete, get, getbypath, insert, list, patch, publish, revert, search and update)

users (get)

Everything else about the blogger API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.posts().insert(...).doit()
let r = hub.posts().publish(...).doit()
let r = hub.posts().delete(...).doit()
let r = hub.posts().revert(...).doit()
let r = hub.posts().update(...).doit()
let r = hub.posts().get_by_path(...).doit()
let r = hub.posts().search(...).doit()
let r = hub.posts().patch(...).doit()
let r = hub.posts().list(...).doit()
let r = hub.posts().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate blogger3;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Blogger::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.posts().list("blogId")
             .view("labore")
             .status("sea")
             .start_date("nonumy")
             .page_token("dolores")
             .order_by("gubergren")
             .max_results(6)
             .labels("aliquyam")
             .fetch_images(false)
             .fetch_bodies(false)
             .end_date("justo")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Blog
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listByUser (none)

get (response)

getByUrl (response)

BlogGetByUrlMethodBuilder
Retrieve a Blog by URL.

BlogGetMethodBuilder
Gets one blog by ID.

BlogList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listByUser (response)

BlogListByUserMethodBuilder
Retrieves a list of blogs, possibly filtered.

BlogLocale
The locale this Blog is set to.

This type is not used in any activity, and only used as part of another schema.

BlogMethodsBuilder
A builder providing access to all methods supported on blog resources. It is not used directly, but through the Blogger hub.

BlogPages
The container of pages in this blog.

This type is not used in any activity, and only used as part of another schema.

BlogPerUserInfo
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

BlogPosts
The container of posts in this blog.

This type is not used in any activity, and only used as part of another schema.

BlogUserInfo
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

BlogUserInfoGetMethodBuilder
Gets one blog and user info pair by blogId and userId.

BlogUserInfoMethodsBuilder
A builder providing access to all methods supported on blogUserInfo resources. It is not used directly, but through the Blogger hub.

Blogger
Central instance to access all Blogger related resource activities

Comment
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listByBlog (none)

get (response)

markAsSpam (response)

list (none)

removeContent (response)

approve (response)

delete (none)

CommentApproveMethodBuilder
Marks a comment as not spam.

CommentAuthor
The author of this Comment.

This type is not used in any activity, and only used as part of another schema.

CommentAuthorImage
The comment creator's avatar.

This type is not used in any activity, and only used as part of another schema.

CommentBlog
Data about the blog containing this comment.

This type is not used in any activity, and only used as part of another schema.

CommentDeleteMethodBuilder
Delete a comment by ID.

CommentGetMethodBuilder
Gets one comment by ID.

CommentInReplyTo
Data about the comment this is in reply to.

This type is not used in any activity, and only used as part of another schema.

CommentList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listByBlog (response)

list (response)

CommentListByBlogMethodBuilder
Retrieves the comments for a blog, across all posts, possibly filtered.

CommentListMethodBuilder
Retrieves the comments for a post, possibly filtered.

CommentMarkAsSpamMethodBuilder
Marks a comment as spam.

CommentMethodsBuilder
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Blogger hub.

CommentPost
Data about the post containing this comment.

This type is not used in any activity, and only used as part of another schema.

CommentRemoveContentMethodBuilder
Removes the content of a comment.

Page
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update (request|response)

get (response)

insert (request|response)

publish (response)

list (none)

patch (request|response)

revert (response)

delete (none)

PageAuthor
The author of this Page.

This type is not used in any activity, and only used as part of another schema.

PageAuthorImage
The page author's avatar.

This type is not used in any activity, and only used as part of another schema.

PageBlog
Data about the blog containing this Page.

This type is not used in any activity, and only used as part of another schema.

PageDeleteMethodBuilder
Delete a page by ID.

PageGetMethodBuilder
Gets one blog page by ID.

PageInsertMethodBuilder
Add a page.

PageList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PageListMethodBuilder
Retrieves the pages for a blog, optionally including non-LIVE statuses.

PageMethodsBuilder
A builder providing access to all methods supported on page resources. It is not used directly, but through the Blogger hub.

PagePatchMethodBuilder
Update a page. This method supports patch semantics.

PagePublishMethodBuilder
Publishes a draft page.

PageRevertMethodBuilder
Revert a published or scheduled page to draft state.

PageUpdateMethodBuilder
Update a page.

PageViewGetMethodBuilder
Retrieve pageview stats for a Blog.

PageViewMethodsBuilder
A builder providing access to all methods supported on pageView resources. It is not used directly, but through the Blogger hub.

Pageviews
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

PageviewsCounts
The container of posts in this blog.

This type is not used in any activity, and only used as part of another schema.

Post
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert (request|response)

publish (response)

delete (none)

revert (response)

update (request|response)

getByPath (response)

search (none)

patch (request|response)

list (none)

get (response)

PostAuthor
The author of this Post.

This type is not used in any activity, and only used as part of another schema.

PostAuthorImage
The Post author's avatar.

This type is not used in any activity, and only used as part of another schema.

PostBlog
Data about the blog containing this Post.

This type is not used in any activity, and only used as part of another schema.

PostDeleteMethodBuilder
Delete a post by ID.

PostGetByPathMethodBuilder
Retrieve a Post by Path.

PostGetMethodBuilder
Get a post by ID.

PostImages
Display image for the Post.

This type is not used in any activity, and only used as part of another schema.

PostInsertMethodBuilder
Add a post.

PostList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

search (response)

PostListMethodBuilder
Retrieves a list of posts, possibly filtered.

PostLocation
The location for geotagged posts.

This type is not used in any activity, and only used as part of another schema.

PostMethodsBuilder
A builder providing access to all methods supported on post resources. It is not used directly, but through the Blogger hub.

PostPatchMethodBuilder
Update a post. This method supports patch semantics.

PostPerUserInfo
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PostPublishMethodBuilder
Publishes a draft post, optionally at the specific time of the given publishDate parameter.

PostReplies
The container of comments on this Post.

This type is not used in any activity, and only used as part of another schema.

PostRevertMethodBuilder
Revert a published or scheduled post to draft state.

PostSearchMethodBuilder
Search for a post.

PostUpdateMethodBuilder
Update a post.

PostUserInfo
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (none)

get (response)

PostUserInfoGetMethodBuilder
Gets one post and user info pair, by post ID and user ID. The post user info contains per-user information about the post, such as access rights, specific to the user.

PostUserInfoListMethodBuilder
Retrieves a list of post and post user info pairs, possibly filtered. The post user info contains per-user information about the post, such as access rights, specific to the user.

PostUserInfoMethodsBuilder
A builder providing access to all methods supported on postUserInfo resources. It is not used directly, but through the Blogger hub.

PostUserInfosList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

User
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

UserBlogs
The container of blogs for this user.

This type is not used in any activity, and only used as part of another schema.

UserGetMethodBuilder
Gets one user by ID.

UserLocale
This user's locale

This type is not used in any activity, and only used as part of another schema.

UserMethodsBuilder
A builder providing access to all methods supported on user resources. It is not used directly, but through the Blogger hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.